add comparator method to inspect subcontents not just string
the expected Authentication header value in the OAuth 1.0a spec example does not place the params in alphabetical order and includes the optional realm param, the new comparator splits the header string into params pairs and compares their contents rather than testing the whole string for equality. perhaps the param ordering is only important to generate the base string for the signature but not in the actual authentication header?